Fixing issues with NativeAVStreamReader and decoding thread
There was an issue in NativeAVStreamReader that was preventing the
filesize to be set so reading and seeking was crashing

Also in the decode thread we removed a condition that check for error
and then put the thread to sleep. This was an issue because an error can
happend but the packet queue isn't empty so all the data available weren't
used (for example reaching EOF would stop the decoding event if the video
is not finished). Instead we rely on processAudio() and processVideo()
if both fails we can put the decode thread to sleep


Former-commit-id: 8adebc6edf9b28cd8a02f7662848fc34c9fdc3d1 [formerly 2e566a65e7fafa9e57361ba0781d76a4dd25b593]
Former-commit-id: ade4257c14fb82728f36067cfb85679e2b26e0eb