Fix nidium/NativeStudio#185 load(); behavior
This commit introduce a change in how load(); works. It also introduce some fix and cleanup :
- Various fix in NativeStream::resolvePath that wasn't consistant
- Remove hardcoded path for framework location (Used by NativeStudio)
- Remove NSS loading handling
- A delegate to override load() behavior

In NativeJSCore the file is loaded relative to the parent script, however a delegate can now override how load() works. In NativeStudio this delegate make load be relative to the NML and also handle NSS loading.