A/V : Fix possible dead lock in queue thread
Under high load the queue thread could constaly stay in the processing loop (because the thread is not processing data fast enought, and thus the output buffer always have space). This behavior is preventing any other thread to access the queue thread lock (for example to add or remove nodes). This commit add a new flag to signal to the queue thread that it needs to give up the processing so other thread can get access to the queue lock.


Former-commit-id: d844f1b0669cf3f51854bde8f3d0614888a048ed [formerly 6b0cad32c37d48b15579978e65d8bae34cb7980f]
Former-commit-id: 4df29c6de686abfc8a3c1d13fe26b1c52f9ecbbc