NativeMessages : Fix to ensure FIFO for postMessage(). Introduce postMessageSync()
With the changes introduced by this commit postMessage() no longer guarantee
that the message will be posted synchronously (even on the same thread) or
asynchronously (unless forceAsync argument is given).

Since various class were relying on a sync posting (for optimisation / ease of
use) this commit introduce a new postMessageSync() method.

All classe that were relying on the old sync behavior of postMessage() has been
updated to use postMessageSync()