Reduced a noticed and make shure that a default db is active.
A default db should be active even if the meta identifier tag is not present in an NML.

Previously with a missing identifier tag, there was a cryptic notice that made the user (un)aware that the db functions were not useable.

It is probably better to init a default database, although this might be a security/privacy issue if developpers are sloppy with their apps.

This makes PR https://github.com/nidium/NativeStudio/pull/341 obsolete.


Former-commit-id: d30433d324104d36ff78e2f6eeb585f4ff57a82b [formerly 8a7c7bca52103d476497281653d4c709bd27786f]
Former-commit-id: ba41b64acf26d10b5a8624cda7e42eb5bd74b871