Skia : Backport GR_GL_PER_GL_FUNC_CALLBACK
In Skia M51 release, the GR_GL_PER_GL_FUNC_CALLBACK is gone and the new way of doing this is by creating a GrGLInterface that wraps all GL call. A sample implementation can be found in mozilla see https://bugzilla.mozilla.org/show_bug.cgi?id=1265131 and https://dxr.mozilla.org/mozilla-central/source/gfx/gl/SkiaGLGlue.cpp
However we are not yet ready to commit time to this task. Instead we reverted the removal of the GR_GL_PER_GL_FUNC_CALLBACK and use this as a temporary workaround