Graphics : Fix internal nidium shader to work with OpenGL ES
This commit introduce a new macro NIDIUM_GL_SHADER_PREAMBLE that has to
be placed before any internal shader code. While we could have directly
wrote  OpenGL ES shader and give the shader to ANGLE this add some
unnecessary extra processing. So instead (and for now) we just use
a simple macro which is enough for our use case.