Elements : Store the tag name as a static property of the class
There wasn't a reliable way to get the tag name of an Element, because
this.constructor.name does not return the correct name in some case and
relying on the override of the name() method was not reliable and force
the user to define a name() method.

Instead we store the name of the tag as a static property of the class.
The name is assigned the first time the tag is created with
Elements.Create(). While this is not 100% safe (the element can still be
created programmatically before his name is set), it still provide an
automated mechanism.

To be completly safe, the alternative is to register new Elements with
a method like Elements.Register(name, class) and set the name at this
point.