windown-x86: setting of dirent types
Problem: with the introduction of the nspr PR_Readdir call, the native file type analysis could not be used anymore.

Please note: it is uncertain if the PR_GetFileInfo call can work on the entry->name. It is well possible that the full pathname must be given.
In that case the onmessage call must be changed to pass the directory name to JSFile_dirtype_to_str.