Android : Fix crash when touchmove event is received before the touchstart event is processed
Touch events are processed when they are dispatched (asynchronously), so
if a touchstart event is received but not yet dispatched/processed and
a touchmove event is received in between, the touchmove event cannot
find it's original Touch object and thus create a new Touch object that
shouldn't exist.

To fix this behavior, the InputHandler class now keep a reference to all
Touch object created (when they are received) and , when a touchmove
event is receievd we can be sure it will find it's original Touch
object.