Context/Events : Fix heap-use-after-free
This fix replace the fix in commit the 8ee6f2c97edc3f4f3f89ad7a593b51c7a893fa6f

The heap-use-after-free was caused when we were accessing the previous
event that has just been deleted. But it's not necessary to keep a reference
to the previous event inside since we are only interested in getting the
CanvasHandler attached to the event.

This commit directly store a reference to the CanvasHandler and thus
suppress the heap-use-after-free.