JSModules : Fix crash caused by dlclose()
When native modules are destroyed, first their export object is garbage
collected and then the actual JSObject of the native module. Calling dlclose()
inside the destructor of the JSModule is making the underlying module
unreachable and thus cause a memory error when SpiderMonkey try to GC it.

To fix this behavior, the call to dlclose() is propagated asynchronously
to the next event loop.