InputHandler : Use "double buffering" for input events list
This commit change the way the linked list of events is handled. On
Android, scrolling events are forwarded to nidium thread through
postMessage() and if another postMessage is made during the propagation
of the event (during/after the call to CanvasHandler::layerize()) the
newly pushed events are discarded before being processed.

To fix the issue, we use two lists for the events :
- One with the events that are currently processed
- One for pending events that will be processed during the next event loop